+++ blocks/form/src/index.tsx
@@ -1,9 +1,10 @@
 import { bootstrap, FormattedMessage } from '@appsemble/preact';
+import type { Remapper } from '@appsemble/sdk';
 import classNames from 'classnames';
 import { h } from 'preact';
 import { useCallback, useEffect, useMemo, useState } from 'preact/hooks';
 
-import type { Field, FileField } from '../block';
+import type { Field, FileField, StringField } from '../block';
 import BooleanInput from './components/BooleanInput';
 import EnumInput from './components/EnumInput';
 import FileInput from './components/FileInput';
@@ -12,8 +13,14 @@ import NumberInput from './components/NumberInput';
 import RadioInput from './components/RadioInput';
 import StringInput from './components/StringInput';
 import styles from './index.css';
+import ValidationError from './utils/ValidationError';
 
-type Validator = (field: Field, event: Event, value: any) => boolean;
+type Validator = (
+  field: Field,
+  event: Event,
+  value: any,
+  remap: (remapper: Remapper, data: any) => any,
+) => boolean;
 
 const inputs = {
   enum: EnumInput,
@@ -27,6 +34,37 @@ const inputs = {
   radio: RadioInput,
 };
 
+const validateString: Validator = (field: StringField, event, value: string, remap) => {
+  const inputValid = (event.target as HTMLInputElement).validity.valid;
+
+  if (!inputValid) {
+    return false;
+  }
+
+  field.requirements?.forEach((requirement) => {
+    let valid = true;
+
+    if ('regex' in requirement) {
+      const regex = new RegExp(requirement.regex, requirement.flags || 'g');
+      valid = regex.test(value);
+    }
+
+    if ('maxLength' in requirement || 'minLength' in requirement) {
+      const maxValid = requirement.maxLength != null ? value.length >= requirement.maxLength : true;
+      const minValid = requirement.minLength != null ? value.length <= requirement.minLength : true;
+
+      valid = maxValid && minValid;
+    }
+
+    if (!valid) {
+      const error = remap(requirement.errorMessage, value);
+      throw new ValidationError(error);
+    }
+  });
+
+  return inputValid;
+};
+
 const validateInput: Validator = (_field, event) =>
   (event.target as HTMLInputElement).validity.valid;
 
@@ -34,6 +72,10 @@ const validators: { [name: string]: Validator } = {
   file: (field: FileField, _event, value) => {
     if (!field.required) {
       return true;
+    }
+
+    if (value === null) {
+      return false;
     }
 
     if (field.accept) {
@@ -51,7 +93,7 @@ const validators: { [name: string]: Validator } = {
   geocoordinates: (_, _event, value: { longitude: number; latitude: number }) =>
     !!(value.latitude && value.longitude),
   hidden: (): boolean => true,
-  string: validateInput,
+  string: validateString,
   number: validateInput,
   integer: validateInput,
   boolean: () => true,
@@ -64,7 +106,7 @@ const messages = {
   unsupported: 'This file type is not supported',
 };
 
-bootstrap(({ actions, data, events, parameters, ready }) => {
+bootstrap(({ actions, data, events, parameters, ready, utils: { remap } }) => {
   const [errors, setErrors] = useState<{ [name: string]: string }>({});
   const [disabled, setDisabled] = useState(true);
   const [validity, setValidity] = useState({
@@ -131,19 +173,34 @@ bootstrap(({ actions, data, events, parameters, ready }) => {
       const field = fields.find((f) => f.name === name);
 
       if (Object.prototype.hasOwnProperty.call(validators, field.type)) {
-        return validators[field.type](field, event, value);
+        return validators[field.type](field, event, value, remap);
       }
       return true;
     },
-    [parameters],
+    [parameters, remap],
   );
 
   const onChange = useCallback(
     (event: Event, value: any): void => {
       const { name } = event.target as HTMLInputElement;
-      const valid = validateField(event, value);
+      let valid: boolean;
+      let error: string;
 
-      setErrors({ ...errors, [name]: valid ? null : 'Invalid' });
+      try {
+        valid = validateField(event, value);
+        if (!valid) {
+          error = messages.invalid;
+        }
+      } catch (e) {
+        if (!(e instanceof ValidationError)) {
+          throw e;
+        }
+
+        valid = false;
+        error = e.message || messages.invalid;
+      }
+
+      setErrors({ ...errors, [name]: error });
       setValidity({ ...validity, [name]: valid });
       setValues({
         ...values,