+++ packages/app/src/components/OpenIDCallback/index.tsx
@@ -1,26 +1,36 @@
-import { Loader, Message, OAuth2LoginButton, useQuery } from '@appsemble/react-components';
+import { Content, Loader, Message, useQuery } from '@appsemble/react-components';
 import { normalize } from '@appsemble/utils';
+import { clearOAuth2State, loadOAuth2State } from '@appsemble/web-utils';
 import * as React from 'react';
 import { FormattedMessage } from 'react-intl';
-import { Redirect } from 'react-router-dom';
+import { Link, Redirect } from 'react-router-dom';
 
-import settings from '../../utils/settings';
 import { useAppDefinition } from '../AppDefinitionProvider';
 import { useUser } from '../UserProvider';
 import styles from './index.css';
 import messages from './messages';
 
+/**
+ * Handle the OAuth2 callback.
+ */
 export default function OpenIDCallback(): React.ReactElement {
-  const { redirect } = localStorage;
   const query = useQuery();
   const code = query.get('code');
+  const errorMessage = query.get('error');
+  const state = query.get('state');
+
+  const session = React.useMemo(() => loadOAuth2State<{}>(), []);
   const { authorizationCodeLogin, isLoggedIn } = useUser();
   const { definition } = useAppDefinition();
 
   const [error, setError] = React.useState(false);
 
+  const { redirect } = session;
+  const stateOk = state && session.state && state === session.state;
+  const isOk = code && !errorMessage && !error && !isLoggedIn && stateOk;
+
   React.useEffect(() => {
-    if (code) {
+    if (isOk) {
       authorizationCodeLogin({
         code,
         redirect_uri: `${window.location.origin}${window.location.pathname}`,
@@ -28,32 +38,35 @@ export default function OpenIDCallback(): React.ReactElement {
         setError(true);
       });
     }
-  }, [authorizationCodeLogin, code]);
+  }, [authorizationCodeLogin, code, isOk]);
 
-  if (!code) {
-    return <FormattedMessage {...messages.invalidCallback} />;
-  }
+  React.useEffect(() => {
+    if (isLoggedIn) {
+      clearOAuth2State();
+    }
+  }, [isLoggedIn]);
 
   if (isLoggedIn) {
     return <Redirect to={redirect || normalize(definition.defaultPage)} />;
   }
 
-  if (error) {
+  if (!isOk) {
     return (
-      <div className={styles.error}>
+      <Content className={styles.error} padding>
         <Message color="danger">
-          <FormattedMessage {...messages.error} />
+          {errorMessage === 'access_denied' ? (
+            <FormattedMessage {...messages.accessDenied} />
+          ) : (
+            <FormattedMessage {...messages.error} />
+          )}
         </Message>
-        <OAuth2LoginButton
-          authorizationUrl={String(new URL('/connect/authorize', settings.apiUrl))}
-          clientId={`app:${settings.id}`}
-          icon="user"
-          redirectUrl="/Callback"
-          scope="openid"
+        <Link
+          className="button"
+          to={{ pathname: '/Login', search: String(new URLSearchParams({ redirect })) }}
         >
           <FormattedMessage {...messages.retry} />
-        </OAuth2LoginButton>
-      </div>
+        </Link>
+      </Content>
     );
   }