+++ packages/server/src/controllers/organizations.test.ts
@@ -33,7 +33,6 @@ beforeEach(async () => {
     name: 'Test Organization',
   });
   await Member.create({ OrganizationId: organization.id, UserId: user.id, role: 'Owner' });
-
   await Organization.create({ id: 'appsemble', name: 'Appsemble' });
 });
 
@@ -177,6 +176,7 @@ describe('getInvites', () => {
 
 describe('inviteMember', () => {
   it('should send an invite to an organization', async () => {
+    await Member.update({ role: 'Maintainer' }, { where: { UserId: user.id } });
     const userB = await EmailAuthorization.create({ email: 'test2@example.com', verified: true });
     await userB.$create('User', { primaryEmail: 'test2@example.com', name: 'John' });
     const response = await request.post(
@@ -191,6 +191,26 @@ describe('inviteMember', () => {
         id: expect.any(String),
         name: 'John',
         primaryEmail: 'test2@example.com',
+      },
+    });
+  });
+
+  it('should not send an invite to an organization if the user does not have the right permissions', async () => {
+    await Member.update({ role: 'AppEditor' }, { where: { UserId: user.id } });
+    const userB = await EmailAuthorization.create({ email: 'test2@example.com', verified: true });
+    await userB.$create('User', { primaryEmail: 'test2@example.com', name: 'John' });
+    const response = await request.post(
+      '/api/organizations/testorganization/invites',
+      { email: 'test2@example.com' },
+      { headers: { authorization } },
+    );
+
+    expect(response).toMatchObject({
+      status: 403,
+      data: {
+        error: 'Forbidden',
+        message: 'User does not have sufficient permissions.',
+        statusCode: 403,
       },
     });
   });
@@ -241,6 +261,7 @@ describe('inviteMember', () => {
 
 describe('resendInvitation', () => {
   it('should resend an invitation', async () => {
+    await Member.update({ role: 'Maintainer' }, { where: { UserId: user.id } });
     const userB = await EmailAuthorization.create({ email: 'test2@example.com', verified: true });
     await userB.$create('User', { primaryEmail: 'test2@example.com', name: 'John' });
 
@@ -257,6 +278,33 @@ describe('resendInvitation', () => {
     );
 
     expect(response).toMatchObject({ status: 204 });
+  });
+
+  it('should not resend an invitation if the user does not have the right permissions', async () => {
+    await Member.update({ role: 'AppEditor' }, { where: { UserId: user.id } });
+    const userB = await EmailAuthorization.create({ email: 'test2@example.com', verified: true });
+    await userB.$create('User', { primaryEmail: 'test2@example.com', name: 'John' });
+
+    await request.post(
+      '/api/organizations/testorganization/invites',
+      { email: 'test2@example.com' },
+      { headers: { authorization } },
+    );
+
+    const response = await request.post(
+      '/api/organizations/testorganization/invites/resend',
+      { email: 'test2@example.com' },
+      { headers: { authorization } },
+    );
+
+    expect(response).toMatchObject({
+      status: 403,
+      data: {
+        error: 'Forbidden',
+        message: 'User does not have sufficient permissions.',
+        statusCode: 403,
+      },
+    });
   });
 
   it('should not resend an invitation to a member who has not been invited', async () => {
@@ -299,6 +347,8 @@ describe('resendInvitation', () => {
 
 describe('removeInvite', () => {
   it('should revoke an invite', async () => {
+    await Member.update({ role: 'Maintainer' }, { where: { UserId: user.id } });
+
     await request.post(
       '/api/organizations/testorganization/invites',
       { email: 'test2@example.com' },
@@ -311,6 +361,30 @@ describe('removeInvite', () => {
     });
 
     expect(response).toMatchObject({ status: 204 });
+  });
+
+  it('should not revoke an invite if the user does not have the right permissions', async () => {
+    await request.post(
+      '/api/organizations/testorganization/invites',
+      { email: 'test2@example.com' },
+      { headers: { authorization } },
+    );
+
+    await Member.update({ role: 'AppEditor' }, { where: { UserId: user.id } });
+
+    const response = await request.delete('/api/organizations/testorganization/invites', {
+      headers: { authorization },
+      data: { email: 'test2@example.com' },
+    });
+
+    expect(response).toMatchObject({
+      status: 403,
+      data: {
+        error: 'Forbidden',
+        message: 'User does not have sufficient permissions.',
+        statusCode: 403,
+      },
+    });
   });
 
   it('should not revoke a non-existent invite', async () => {