+++ packages/studio/src/components/BlockList/index.tsx
@@ -1,6 +1,5 @@
-import { Loader, Message } from '@appsemble/react-components';
+import { Loader, Message, useData } from '@appsemble/react-components';
 import type { BlockManifest } from '@appsemble/types';
-import axios from 'axios';
 import React from 'react';
 import { FormattedMessage } from 'react-intl';
 
@@ -13,17 +12,7 @@ import messages from './messages';
  * Display a list of cards representing the available blocks.
  */
 export default function BlockList(): React.ReactElement {
-  const [blocks, setBlocks] = React.useState<BlockManifest[]>();
-  const [loading, setLoading] = React.useState(true);
-  const [error, setError] = React.useState(false);
-
-  React.useEffect(() => {
-    axios
-      .get<BlockManifest[]>('/api/blocks')
-      .then((result) => setBlocks(result.data.sort((a, b) => a.name.localeCompare(b.name))))
-      .catch(() => setError(true))
-      .finally(() => setLoading(false));
-  }, []);
+  const { data: blocks, error, loading } = useData<BlockManifest[]>('/api/blocks');
 
   if (error) {
     return (
@@ -37,8 +26,12 @@ export default function BlockList(): React.ReactElement {
     return <Loader />;
   }
 
-  const appsembleBlocks = blocks.filter((b) => b.name.startsWith('@appsemble'));
-  const thirdPartyBlocks = blocks.filter((b) => !b.name.startsWith('@appsemble'));
+  const appsembleBlocks = blocks
+    .filter((b) => b.name.startsWith('@appsemble'))
+    .sort((a, b) => a.name.localeCompare(b.name));
+  const thirdPartyBlocks = blocks
+    .filter((b) => !b.name.startsWith('@appsemble'))
+    .sort((a, b) => a.name.localeCompare(b.name));
 
   return (
     <>