+++ packages/utils/src/api/paths/emails.ts
@@ -13,6 +13,9 @@ export default {
               type: 'object',
               required: ['email', 'password'],
               properties: {
+                name: {
+                  type: 'string',
+                },
                 email: {
                   type: 'string',
                   format: 'email',