+++ CHANGELOG.md
@@ -10,7 +10,14 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ### Added
 
 - Block(`form`): Add `submitLabel` property for customizing the label on the submit button.
+- Block(`form`): Add support for `requirements` in string fields. Please consult the block
+  documentation for further details.
 - Block(`form`): Add support for `radio` type fields.
+- Studio: Add new `InviteMember` permission to the maintainer role within an organization. They are
+  allowed to invite members, resend invitations and delete pending invitations.
+- Studio: Add `Default` column to block documentation.
+- Server: Support optional display name upon registration.
+- Studio: Support optional display name upon registration.
 
 ## [0.13.3] - 2020-06-11